requirements.txt
streamlit>=1.31.0
numpy>=1.24.0
pandas>=2.0.0
duckdb>=0.9.0
pytz>=2023.3
azure-storage-blob>=12.18.0
configparser>=6.0.0
openpyxl>=3.1.2
python-dateutil>=2.8.2
